Fix sidebar and routing issues
