+++ web/src/app/(main)/layout.tsx
@@ -1,4 +1,5 @@
 import BottomNav from "@/components/layout/BottomNav";
+import Sidebar from "@/components/layout/Sidebar";
 
 export default function MainLayout({
   children,
@@ -6,8 +7,9 @@ export default function MainLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <div className="min-h-screen pb-16 md:pb-0">
-      <main>{children}</main>
+    <div className="min-h-screen">
+      <Sidebar />
+      <main className="pb-16 md:pb-0 md:ml-64">{children}</main>
       <BottomNav />
     </div>
   );
